User name must be email.
Too complicated to allow for a user name and an email. Most services require email only. In the future this will make a "no leak" mode easier to implement in which we do not leak if an account exists already. This is a to-do. Also makes dealing with potential duplicate user names easier and less likely.